refactor: move components into layout and ui folders
