+++ src/components/layout/RootLayout.tsx
@@ -1,8 +1,8 @@
 import Image from "next/image"
 import Link from "next/link"
-import Header from "@components/Header/Header"
-import Footer from "@components/Footer"
-import Main from "@components/Main"
+import Header from "@components/layout/Header/Header"
+import Footer from "@/components/layout/Footer"
+import Main from "@/components/layout/Main"
 
 export default function RootLayout({
   children,
