+++ src/pages/index.tsx
@@ -1,5 +1,3 @@
-import Card from "@components/Card"
-
 export default function Home() {
   return <h1 className="text-xl">Home page</h1>
 }
